# search/internal/rest/api.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Search",
    "version": "1.0.0"
  },
  "tags": [
    {
      "name": "SearchService"
    }
  ],
  "basePath": "/",
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
    "/api/search/bootstrap/demo": {
      "get": {
        "operationId": "SearchService_GetDemoBootstrap",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/searchpbGetDemoBootstrapResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "tags": [
          "SearchService"
        ]
      }
    },
    "/api/search/orders": {
      "post": {
        "summary": "Search for orders",
        "operationId": "searchOrders",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/searchpbSearchOrdersResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/searchpbSearchOrdersRequest"
            }
          }
        ],
        "tags": [
          "Orders"
        ]
      }
    },
    "/api/search/orders/{id}": {
      "get": {
        "summary": "Get an order",
        "operationId": "getOrder",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/searchpbGetOrderResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/rpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "Order"
        ]
      }
    }
  },
  "definitions": {
    "OrderItem": {
      "type": "object",
      "properties": {
        "productId": {
          "type": "string"
        },
        "storeId": {
          "type": "string"
        },
        "productName": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "price": {
          "type": "number",
          "format": "double"
        },
        "quantity": {
          "type": "string",
          "format": "int64"
        }
      }
    },
    "SearchOrdersRequestFilters": {
      "type": "object",
      "properties": {
        "customerId": {
          "type": "string"
        },
        "after": {
          "type": "string",
          "format": "date-time"
        },
        "before": {
          "type": "string",
          "format": "date-time"
        },
        "storeIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "productIds": {
          "type": "array",
          "items": {
            "type": "string"
          }
        },
        "minTotal": {
          "type": "number",
          "format": "double"
        },
        "maxTotal": {
          "type": "number",
          "format": "double"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
    "rpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "searchpbDemoCustomer": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "searchpbDemoProduct": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "storeId": {
          "type": "string"
        },
        "storeName": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "searchpbDemoStore": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "name": {
          "type": "string"
        }
      }
    },
    "searchpbGetDemoBootstrapResponse": {
      "type": "object",
      "properties": {
        "customer": {
          "$ref": "#/definitions/searchpbDemoCustomer"
        },
        "stores": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/searchpbDemoStore"
          }
        },
        "products": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/searchpbDemoProduct"
          }
        }
      }
    },
    "searchpbGetOrderResponse": {
      "type": "object",
      "properties": {
        "order": {
          "$ref": "#/definitions/searchpbOrder"
        }
      }
    },
    "searchpbOrder": {
      "type": "object",
      "properties": {
        "orderId": {
          "type": "string"
        },
        "customerId": {
          "type": "string"
        },
        "customerName": {
          "type": "string"
        },
        "items": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/OrderItem"
          }
        },
        "total": {
          "type": "number",
          "format": "double"
        },
        "status": {
          "type": "string"
        }
      }
    },
    "searchpbSearchOrdersRequest": {
      "type": "object",
      "properties": {
        "filters": {
          "$ref": "#/definitions/SearchOrdersRequestFilters"
        },
        "next": {
          "type": "string"
        },
        "limit": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "searchpbSearchOrdersResponse": {
      "type": "object",
      "properties": {
        "orders": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/searchpbOrder"
          }
        },
        "next": {
          "type": "string"
        }
      }
    }
  }
}